refactor: split backend routes per page
